:sparkles: add options missing columns

+++ database/migrations/2020_11_28_011112_create_adicionais_table.php
@@ -17,8 +17,8 @@ class CreateAdicionaisTable extends Migration
             $table->id();
             $table->integer("produto_id");
             $table->string("titulo");
-            $table->float("maximo");
-            $table->float("minimo");
+            $table->integer("maximo");
+            $table->integer("minimo");
             $table->timestamps();
         });
     }
@@ -33,5 +33,3 @@ class CreateAdicionaisTable extends Migration
         Schema::dropIfExists("adicionais");
     }
 }
-
-
